fix: issues regarding session queries
